+++ lib/c_mini/src/assign.mli
@@ -33,6 +33,10 @@ val rvalue : t -> Expression.t
 module On_addresses :
   Travesty.Traversable_types.S0 with type t = t and type Elt.t = Address.t
 
+(** Traversing over expressions in assignments. *)
+module On_expressions :
+  Travesty.Traversable_types.S0 with type t = t and type Elt.t = Expression.t
+
 (** Traversing over lvalues in assignments. *)
 module On_lvalues :
   Travesty.Traversable_types.S0 with type t = t and type Elt.t = Lvalue.t