+++ lib/c_mini/src/expression.ml
@@ -110,6 +110,11 @@ module Make_traversal (Basic : sig
       with type t := Address.t
        and module Elt = Elt
 
+  module C :
+    Travesty.Traversable_types.S0
+      with type t := Constant.t
+       and module Elt = Elt
+
   module L :
     Travesty.Traversable_types.S0
       with type t := Atomic_load.t
@@ -123,11 +128,12 @@ Travesty.Traversable.Make0 (struct
   module On_monad (M : Monad.S) = struct
     module F = Travesty.Traversable.Helpers (M)
     module A = Basic.A.On_monad (M)
+    module C = Basic.C.On_monad (M)
     module L = Basic.L.On_monad (M)
 
     let rec map_m x ~f =
       Variants.map x
-        ~constant:(F.proc_variant1 M.return)
+        ~constant:(F.proc_variant1 (C.map_m ~f))
         ~address:(F.proc_variant1 (A.map_m ~f))
         ~uop:
           (F.proc_variant2 (fun (u, x) ->
@@ -144,13 +150,36 @@ Travesty.Traversable.Make0 (struct
   end
 end)
 
+(* TODO(@MattWindsor91): travesty *)
+module Ignore (T : T) (Elt : Equal.S) = Travesty.Traversable.Make0 (struct
+  type t = T.t
+
+  module Elt = Elt
+
+  module On_monad (M : Monad.S) = struct
+    let map_m (x : t) ~(f : Elt.t -> Elt.t M.t) : t M.t =
+      ignore f ; M.return x
+  end
+end)
+
 module On_addresses :
   Travesty.Traversable_types.S0 with type t = t and type Elt.t = Address.t =
 Make_traversal (struct
   module Elt = Address
   module A =
     Travesty.Traversable.Fix_elt (Travesty_containers.Singleton) (Address)
+  module C = Ignore (Constant) (Address)
   module L = Atomic_load.On_addresses
+end)
+
+module On_constants :
+  Travesty.Traversable_types.S0 with type t = t and type Elt.t = Constant.t =
+Make_traversal (struct
+  module Elt = Constant
+  module A = Ignore (Address) (Constant)
+  module C =
+    Travesty.Traversable.Fix_elt (Travesty_containers.Singleton) (Constant)
+  module L = Ignore (Atomic_load) (Constant)
 end)
 
 module On_lvalues :
@@ -158,12 +187,9 @@ module On_lvalues :
 Make_traversal (struct
   module Elt = Lvalue
   module A = Address.On_lvalues
+  module C = Ignore (Constant) (Lvalue)
   module L = Atomic_load.On_lvalues
 end)
-
-module On_identifiers :
-  Travesty.Traversable_types.S0 with type t = t and type Elt.t = Identifier.t =
-  Travesty.Traversable.Chain0 (On_lvalues) (Lvalue.On_identifiers)
 
 module Type_check (E : Env_types.S) = struct
   module Ad = Address.Type_check (E)