+++ lib/c_mini/src/statement_types.mli
@@ -34,6 +34,12 @@ module type S_common = sig
         with type t = t
          and type Elt.t = Address.t
 
+    (** Traversing over expressions. *)
+    module On_expressions :
+      Travesty.Traversable_types.S0
+        with type t = t
+         and type Elt.t = Expression.t
+
     (** Traversing over lvalues. *)
     module On_lvalues :
       Travesty.Traversable_types.S0 with type t = t and type Elt.t = Lvalue.t